[serlib] New serialization import method.
We make a better use of module layering to avoid having to use
`[@with t := nt]` clauses in our `ppx_import` declarations.

This saves much typing and is, in general, more robust. Note that we
depend on the module aliases feature as not to blow-up the code size.

The general idea is that the corresponding serialization module defines
internally the same modules than the corresponding Coq module, thus, at
ppx processing time, all that we have to do is:

```ocaml
module T = Ser_t
```

so the right `T.t_of_sexp` functions become available on import.