[serlib] More support for generic arguments.
Working proof of concept to serialize values. It works, however it
illustrates some layering problems as genarg cannot directly access
the upper serializers.

Coq uses a registry to avoid this problem in printing and other
places, we must think more about what is the desired solution for us.
An alternative to the low-level registry is indeed a top-level
delegation.

For the moment, we sketch a possible entry type for generic argument
serializers.